fix homepage swiper
